Simplified (and fixed) the graph.
There was a bug because the values where in actuals and I just
converted the y axis labels, but that only works for the reference
country and the multiples of the plotted country become distorted.
Now I calculate the multiples for both countries instead and the
y axis labels sort themselves out easier as well.